+++ lcs/agents/racs/Configuration.py
@@ -9,6 +9,9 @@ class Configuration:
                  encoder_bits: int,
                  perception_mapper_fcn=None,
                  action_mapping_fcn=None,
+                 environment_metrics_fcn=None,
+                 performance_fcn=None,
+                 performance_fcn_params={},
                  do_ga=False,
                  do_subsumption=True,
                  beta=0.05,
@@ -32,6 +35,10 @@ class Configuration:
 
         self.perception_mapper_fcn = perception_mapper_fcn
         self.action_mapping_fcn = action_mapping_fcn
+        self.environment_metrics_fcn = environment_metrics_fcn
+        self.performance_fcn = performance_fcn
+        self.performance_fcn_params = performance_fcn_params
+
         self.do_ga = do_ga
         self.do_subsumption = do_subsumption